Added new rem mixin. Converted to being relative to base font size rather than hard coded 10px value. Closes #143
Discovered that using `62.5%` trick is a bit outdated and bad for
accessibility as you're forcing a illegibly small root font size on a
user and then having to re-declare it to bump it to a useable size on
the body.

Revised approach so that rem mixin is relative to whatever developer
declares as baseFontSize rather than hard coded 10px value. Now html is
set to 100% and body is set in rems to a relative value equivilant to
16px (assuming user has not altered browser default text size
settings). If user alters text setting defaults then mixin still
functions correctly in a "relative" manner.